backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI Backend for audio transcription service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.116.1",
    "langchain>=1.2.0,<2",
    "langchain-anthropic>=0.3.0,<1",
    "langchain-google-genai>=2.1.0,<3",
    "langgraph>=1.1.8,<2",
    "langchain-openai>=0.3.0,<1",
    "httpx==0.28.1",
    "pydantic-settings==2.10.1",
    "uvicorn[standard]==0.35.0",
    "websockets==15.0.1",
]
